Now uses attach instead of accepting specs straight away
There were two sets of modifiers: The specs one and the DSL's one.
Because it wasn't clear how to treat them, it is changed to a clear way.

The old way of adding was just an extension function.
This means that the old way can be added by implementing the methods from own code.